Improve the markov generator for words
I changed the way that it generates words. I now have the network include
nodes that say when the words end. This changes the way the network is
trained, but does not change the interface. It also changes the generation
method.

I also added a script for creating a shorter random word list from the
master list. This is meant to allow better training tests. However this
script is not deterministic.

Removed the uppercasing the words from the filters that generate the master
list. This makes it more realistic for real world data.